Draw pictures for ENM results

主要用于绘制适生区的图, 例如maxent生成的连续型适宜性栅格, 重分类后的分类适宜性栅格等. 预设一套背景模版, 可以在该模版上替换不同的栅格和矢量绘制常用的地图. 也可以结合ggplot2及其扩展绘制更个性化的地图.

Usage

ENMplots(
  spatraster = NULL,
  spatvector = NULL,
  zones = NULL,
  expand = 1,
  rastercolors = NULL,
  landcolor = "white",
  watercolor = "#BEE8FF",
  maxcell = 5e+05,
  categories = F,
  crs = NULL,
  croptoraster = T,
  annotation_north_arrow = T,
  annotation_scale = T,
  arrange = list(nrow = 2, ncol = 2, dir = "v"),
  width,
  height,
  device = "jpg",
  filename,
  outdir
)

Arguments

spatraster

spatraster对象, 通常为物种分布的适宜性栅格.

spatvector

spatvector对象, 通常为研究区的矢量地图.

zones

NULL or SpatRaster with the same geometry identifying zones in spatraster.

expand

数值向量, 大于等于1, 长度为1或4. 按比例扩展绘图范围. 当长度为4时分别对应左右下上.

rastercolors

字符型, 创建渐变色(continuous scale)或分类色(discrete scale)用于设置适宜性栅格的颜色.

landcolor

除研究区外的其他陆地颜色. 用来做背景色.

watercolor

缺失值的颜色, 通常代表水体颜色.

maxcell

positive integer. Maximum number of cells to use for the plot.

categories

逻辑值. 将spatraster视为类别型还是连续型栅格.

crs

The coordinate reference system (CRS) into which all data should be projected before plotting. If not specified, will use the CRS defined in the first sf layer of the plot.

croptoraster

逻辑值. 是否将spatvector裁剪至spatraster范围.

arrange

包含3个元素的列表, 分别为nrow, ncol, dir. 当spatraster为多层时, 将使用分面绘制每层栅格, nrow和ncol 设置一张图片上分图的行列数, dir设置分图的排列方向, dir = "v"则按水平方向排列, dir = "h"则按垂直方向排列. 如果层数大于nrow* ncol, 则会生成多张图片, 直到绘制完所有层.

width

图片宽度(cm).

height

图片高度(cm).

filename

图片名称.

outdir

图片保存路径.

Value

ggplot2对象

Examples

# 单层栅格绘图
## 下载澳大利亚行政边界数据
aus_map <- geodata::gadm(country = "AUS", level = 1, path = tempdir())
#> Error in loadNamespace(x): there is no package called ‘geodata’
## 下载全球12个月的平均温度栅格数据
tavg <- geodata::worldclim_global("tavg", res = 10, path = tempdir(), version = "2.1")
#> Error in loadNamespace(x): there is no package called ‘geodata’
## 裁剪至澳大利亚范围
tavg_aus <- terra::crop(tavg, aus_map, mask = T)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'crop': object 'tavg' not found
## 提取1月的数据
tavg_aus_Jan <- tavg_aus[[1]]
#> Error: object 'tavg_aus' not found
## 绘图
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map
)
#> Error: object 'tavg_aus_Jan' not found
## 更改陆地的颜色和水体的颜色
### 如果不需要陆地和水体颜色可以将颜色设置为"transparent", 透明色.
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  landcolor = "grey60",
  watercolor = "#BEE8FF"
)
#> Error: object 'tavg_aus_Jan' not found
## 自定义温度栅格的颜色
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  rastercolors = c("yellow", "green")
)
#> Error: object 'tavg_aus_Jan' not found
## 移除指北针和比例尺
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  annotation_north_arrow = F,
  annotation_scale = F
)
#> Error: object 'tavg_aus_Jan' not found
## 更改投影坐标系
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  crs = "epsg:4544"
)
#> Error: object 'tavg_aus_Jan' not found
## 仅绘制区域地图
### 区域地图需要提前准备一个与spatraster具有相同地理参考的栅格数据. 如果只想展示研究区的一部分结果可能很有用, 例如某个省.
zone <- terra::crop(tavg_aus_Jan, terra::ext(tavg_aus_Jan) + c(-27, -5, -10, -30))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'crop': object 'tavg_aus_Jan' not found
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  croptoraster = F,
  zones = zone
)
#> Error: object 'tavg_aus_Jan' not found
### 指定croptoraster = T将绘图范围裁剪至zone
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  croptoraster = T,
  zones = zone
)
#> Error: object 'tavg_aus_Jan' not found
## 保存图片
### 可以在函数中使用参数保存图片, 请在filename中指定图片格式. 支持的格式同ggplot2::ggsave().
ENMplots(
  spatraster = tavg_aus_Jan,
  spatvector = aus_map,
  outdir = "your path",
  filename = "plot.jpg",
  width = 10,
  height = 10
)
#> Error: object 'tavg_aus_Jan' not found

## 添加其他ggplot对象
### ENMplots()返回的是ggplot对象，可以与其他ggplot扩展结合，拥有更大的灵活性
### 使用tidyverse包的调色盘并更换主题
ENMplots(spatraster = tavg_aus_Jan, spatvector = aus_map) +
  tidyterra::scale_fill_whitebox_c(
    palette = "high_relief",
    labels = scales::label_number(suffix = "º"),
    n.breaks = 12,
    guide = guide_legend(reverse = TRUE)
  ) +
  theme_bw()
#> Error: object 'tavg_aus_Jan' not found
### 如果添加了额外的ggplot对象，则需要使用ggplot2::ggsave()在外部保存图片.

## 此外还可以绘制简单的样本分布点
### 随机选择100个点当做样本分布点
point <- terra::spatSample(tavg_aus_Jan, 100, xy = T, na.rm = T)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'spatSample': object 'tavg_aus_Jan' not found
ENMplots(
  spatraster = NULL, # 如果不需要温度栅格就设置为NULL
  spatvector = aus_map
) +
  ggplot2::geom_point(data = point, aes(x = x, y = y))
#> Error: object 'aus_map' not found
##单层分类栅格绘图
##将澳大利亚1月的平均温重分类为高中低(3, 2, 1)三个类别
rcl <- matrix(c(0,20,1,20,30,2,30,35,3), ncol = 3, byrow = TRUE)
tavg_aus_Jan_class <- terra::classify(tavg_aus_Jan, rcl = rcl)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'classify': object 'tavg_aus_Jan' not found
##绘图
###如果想要对分类栅格绘图，需要设置categories=T. 同样, 如果想要对连续栅格绘图需要设置categories=F.
ENMplots(spatraster = tavg_aus_Jan_class,
         spatvector = aus_map,
         categories = T)
#> Error: object 'tavg_aus_Jan_class' not found
# 多层栅格绘图
##多层栅格绘图主要用于批量绘图并对图片排版
###首先需要创建含有多层的spatraster, tavg_aus包含了12个层(12个月的平均温度).
ENMplots(spatraster = tavg_aus,
         spatvector = aus_map,
         categories = F,
         arrange = list(2, 3, "v"), #使用arrange参数设置每张图片上各个分图的数量及排列方式.
         outdir = "your path",
         device = "pdf",
         width = 20,
         height = 20)
#> Error: object 'tavg_aus' not found
####注意: 当多层绘图时不用对生成的图片命名, 内部会自动对每张图进行编号. 需要通过device参数设置图片格式.